Added if(Yii::->user->can()) to all actions in PictureController

+++ frontend/controllers/PictureController.php
@@ -6,6 +6,7 @@ use Yii;
 use frontend\models\Picture;
 use yii\data\ActiveDataProvider;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -32,13 +33,18 @@ class PictureController extends Controller
      */
     public function actionIndex()
     {
-        $dataProvider = new ActiveDataProvider([
-            'query' => Picture::find(),
-        ]);
+        if( Yii::$app->user->can('picture-index') ) {
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-        ]);
+            $dataProvider = new ActiveDataProvider([
+                'query' => Picture::find(),
+            ]);
+
+            return $this->render('index', [
+                'dataProvider' => $dataProvider,
+            ]);
+        }else {
+            throw new ForbiddenHttpException;
+        }
     }
 
     /**
@@ -48,9 +54,13 @@ class PictureController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if( Yii::$app->user->can('picture-view') ) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }else {
+            throw new ForbiddenHttpException;
+        }
     }
 
     /**
@@ -60,55 +70,56 @@ class PictureController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Picture();
+        if( Yii::$app->user->can('picture-create') ) {
+            $model = new Picture();
 
-        if( isset($_GET['id']) ) {
-
-            $model->person_id = (int)$_GET['id'];
-        }
-
-
-        if ($model->load(Yii::$app->request->post()) ) {
-
-            $path = $model->createImageFolder();
-
-            $randHex = substr(md5(rand()), 0, 8);
-
-            $model->file_name = $path . $randHex;
-
-            $file = \yii\web\UploadedFile::getInstance($model, 'file_name');
-
-            $pathToImageFileFullSize = 'pictures/' . $model->file_name . ".jpg";
-
-            $pathToImageFileSmallSize = 'pictures/' . $model->file_name . "_t.jpg";
-
-            $file->saveAs($pathToImageFileFullSize);
-
-            copy($pathToImageFileFullSize, $pathToImageFileSmallSize);
-
-            $model->reduceToFullSize($pathToImageFileFullSize);
-
-            $model->reduceToSmallSize($pathToImageFileSmallSize);
-
-
-            /*
-             * if avatar was selected, remove any avatar previously selected
-             */
-            if($model->avatar == 1)
-            {
-                $model->setAvatarToZeroByPersonId($model->person_id);
-                $model->save();
-                return $this->redirect(['person/profile', 'id' => $model->person_id]);
-
-            } else
-            {
-                $model->save();
-                return $this->redirect(['person/portfolio', 'id' => $model->person_id]);
+            if( isset($_GET['id']) ) {
+                $model->person_id = (int)$_GET['id'];
             }
 
-        } else {
-            return $this->render('create', ['model' => $model,]);
+            if ($model->load(Yii::$app->request->post()) ) {
+
+                $path = $model->createImageFolder();
+
+                $randHex = substr(md5(rand()), 0, 8);
+
+                $model->file_name = $path . $randHex;
+
+                $file = \yii\web\UploadedFile::getInstance($model, 'file_name');
+
+                $pathToImageFileFullSize = 'pictures/' . $model->file_name . ".jpg";
+
+                $pathToImageFileSmallSize = 'pictures/' . $model->file_name . "_t.jpg";
+
+                $file->saveAs($pathToImageFileFullSize);
+
+                copy($pathToImageFileFullSize, $pathToImageFileSmallSize);
+
+                $model->reduceToFullSize($pathToImageFileFullSize);
+
+                $model->reduceToSmallSize($pathToImageFileSmallSize);
+
+
+                /*
+                 * if avatar was selected, remove any avatar previously selected
+                 */
+                if($model->avatar == 1) {
+                    $model->setAvatarToZeroByPersonId($model->person_id);
+                    $model->save();
+                    return $this->redirect(['person/profile', 'id' => $model->person_id]);
+                } else {
+                    $model->save();
+                    return $this->redirect(['person/portfolio', 'id' => $model->person_id]);
+                }
+
+            } else {
+                return $this->render('create', ['model' => $model,]);
+            }
+
+        }else {
+            throw new ForbiddenHttpException;
         }
+
     }
 
     /**
@@ -119,25 +130,31 @@ class PictureController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if( Yii::$app->user->can('picture-update') ) {
 
-        if ($model->load(Yii::$app->request->post())) {
+            $model = $this->findModel($id);
 
-            if ($model->avatar) {
-                $model->setAvatarToZeroByPersonId($model->person_id);
+            if ($model->load(Yii::$app->request->post())) {
+
+                if ($model->avatar) {
+                    $model->setAvatarToZeroByPersonId($model->person_id);
+                }
+
+                $model->save();
+
+                if ($model->avatar) {
+                    return $this->redirect(['person/profile', 'id' => $model->person_id]);
+                }
+
+                return $this->redirect(['person/select', 'id' => $model->person_id]);
+            } else {
+
+                return $this->render('update', ['model' => $model,]);
             }
-
-            $model->save();
-
-            if ($model->avatar) {
-                return $this->redirect(['person/profile', 'id' => $model->person_id]);
-            }
-
-            return $this->redirect(['person/select', 'id' => $model->person_id]);
-        } else {
-
-            return $this->render('update', ['model' => $model,]);
+        }else {
+            throw new ForbiddenHttpException;
         }
+
     }
 
     /**
@@ -148,19 +165,30 @@ class PictureController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if( Yii::$app->user->can('picture-delete') ) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+
+        }else {
+            throw new ForbiddenHttpException;
+        }
     }
 
 
     public function actionDisplay($id)
     {
-        $model = new Picture();
+        if( Yii::$app->user->can('picture-display') ) {
 
-        return $this->render('display', ['model' => $this->findModel($id),
-            'previous' => $model->getPreviousPicture($id),
-            'next' => $model->getNextPicture($id) ]);
+            $model = new Picture();
+
+            return $this->render('display', ['model' => $this->findModel($id),
+                'previous' => $model->getPreviousPicture($id),
+                'next' => $model->getNextPicture($id) ]);
+        }else {
+            throw new ForbiddenHttpException;
+        }
+
     }
 
     /**
